complete2

+++ src/main/java/se331/lab/dao/EventDao.java
@@ -6,6 +6,6 @@ import org.springframework.data.domain.Page;
 public interface EventDao {
     Integer getEventSize();
     Page<Event> getEvents(Integer pageSize, Integer page);
-    Event getEvent(Integer id);
+    Event getEvent(Long id);
     Event save(Event event);
 }
