+++ src/main/java/se331/lab/service/OrganizerServiceImpl.java
@@ -6,8 +6,6 @@ import org.springframework.stereotype.Service;
 import se331.lab.dao.OrganizerDao;
 import se331.lab.entity.Organizer;
 
-import java.util.List;
-
 @Service
 @RequiredArgsConstructor
 public class OrganizerServiceImpl implements OrganizerService {
@@ -24,7 +22,7 @@ public class OrganizerServiceImpl implements OrganizerService {
     }
 
     @Override
-    public Organizer getOrganizer(Integer id) {
+    public Organizer getOrganizer(Long id) {
         return organizerDao.getOrganizer(id);
     }
